# feat: Add high-performance Go implementation for Claude session message search