… all_comparison.json
… all_comparison.md
… fullparse_comparison.json
… fullparse_comparison.md
… fullparse_opt_vs_rust.json
… fullparse_opt_vs_rust.md
… golang/BENCHMARK_RESULTS.md
… golang/README.md
… golang/cmd/benchmark/main.go
… golang/cmd/benchmark/simple_main.go
… golang/cmd/search_fast/main.go
… golang/cmd/search_final/main.go
… golang/cmd/search_fullparse/main.go
… golang/cmd/search_fullparse_optimized/main.go
… golang/cmd/search_jsonparser_optimized/main.go
… golang/cmd/search_optimized/main.go
… golang/cmd/search_profile/main.go
… golang/cmd/search_simd/main.go
… golang/cmd/search_simdjson/main.go
… golang/cmd/search_sonic/main.go
… golang/cmd/search_ultra/main.go
… golang/cmd/search_ultra_safe/main.go
… golang/go.mod
… golang/go.sum
… golang/internal/schemas/session_message.go
… golang/internal/schemas/simple_message.go
… golang/internal/search/loader.go
… golang/internal/search/search.go
… golang/internal/search/simple_loader.go
… golang/internal/search/simple_search.go